Authentication scheme selector enhancements.
The authentication scheme selector is now implemented as an AuthenticationSchemeProvider.  This works out better for parsing the auth tokens and then dispatching to the correct authentication scheme.

Also the logic that determines which authN scheme to use is now configured by a list of delegates that is configured at program startup.  This will make it easier to add/remove authN schemes for now projects.